Create ch1_prog_7_smoothed_line_1.py
Objective: Draw a line with a smooth bend on a canvas. 

Comments: When drawing lines with many segments you specify location of 
each segment. The end of one segment is the start of the next. 
In addition choosing the option "smooth="true" gets the line smoothed into  a curve. 
The amount of smoothness in the curve is determined by the number of spline steps. 
High numbers give a smoother curve. 
In this example the default number of splinesteps is in effect.