Create ch2_prog_3_radius_center_circle_1.py
Objective: A circle specified in terms of its radius and center position. 

Comments:A circle is a special case of an oval and is defined by the box it fits inside. The bounding box here is defined in terms of the radius and position of the center. This is a more convenient and logical  way of drawing a circle (geometrically speaking).