Create ch2_prog_4_concentric_circles_1.py
Objective: A set of circles specified in terms of their radius and center position. 

Comments: A circle is a special case of an oval and is defined by the box it fits inside. The bounding box here is defined in terms of the radius and position of the center. This is a more convenient and logical  way of drawing a circle (geometrically speaking).