Create ch2_prog_5_loop_generated_concentric_circles_3.py
Objective: Use a for loop to generate a large set of concentric circles . 

Comments: A circle is a special case of an oval and is defined by the box it fits inside. The bounding box here is defined in terms of the radius and position of the center. This is a more convenient and logical  way of drawing a circle (geometrically speaking).