Create ch4_prog_4_positioned_sized_star_1.py
Objective: Draw a series of stars each with their own start position and size (scaled larger or smaller from a prototype). 

Comments: A scale factor "size_scaling" is used to change the size of the stars.
By having each separate star draw itself relative to a pair variables, 
x_anchor and y_anchor as well as the size_scaling, it simple to re-position and re-size different stars just by reassigning the values of the anchor position.