Create ch4_prog_7_poly_python_1.py
Objective: Illustrate polygon shape parameters. 

Comments: Take note of the instruction  "for i in range(int(len(shape)/2)):" in the function "locate_shape(shape)". It has to have the "int(...) in order to work under Python 3.xx. 
Without it we get and error as follows: 

  for i in range(len(shape)/2):    # Separate out the x and components. 
TypeError: 'float' object cannot be interpreted as an integer 

Under Python 2.xx the form "for i in range(len(shape)/2):" works fine.