Create ch5_prog_9_line_rotation_randomize_1.py
Objective: Add random components to a line between two points and then rotate the line through a full circle. 

Comments:A Straight line between two points has a random 'wobble' added to it. 
Two styles of randomization are used. The leftmost is random in both vertical and horizontal components. The rightmost is only random in the vertical component.