Create ch5_prog_10_line_rough_randomization_1.py
Objective: Add jagged random components to a line between two points. 

Comments: A Straight line between two points has a random 'wobble' added to it. 
To confine the randomness to being deviations on either side of the line without overlap or looping a strategy or rotation to the x-axis, 
randomization and restore the rotation was used.