Create ch6_prog_4_text_font_space_1.py
Objective: Draw text onto the canvas at a chosen location. 

Comments: The text is written starting at co-ordinate location x= 200 with y progressing down the canvas in steps of 30 pixels. 
The default position has the text string centered at x= 200 which can look messy. 
The "width=500" is the horizontal size of the writing area, in pixels (not typeset point measure).