Create ch6_prog_9_text_justify_align_1.py
Objective: Place different colored texts and font types on a canvas. 

Comments: 
1. Font type, size and color can be specified. Tkinter will apply a 'best-guess' font if the exact font specified is not available 
on your platform. 

2. anchor=anchorPos:  Text position orig_x, orig_y is where the text is centered (by default). 
   This can be overriden by anchor options like "anchor= NW". 

anchorPos Specifies which part of the text (the text's bounding box, more exactly) that should be placed at the given position. 
Use one of N, NE, E, SE, S, SW, W, NW or CENTER. Default is CENTER. 

3. justify=how (ACTIVE WHEN LINE OVERFLOWS OCCUR) 
Specifies how to justify the text within its bounding region. How must be one of the values "LEFT", "RIGHT", or "CENTER". This option will only matter if the text is displayed as multiple lines. If the option is omitted, it defaults to left. 

4. tags=tagList 
Specifies a set of tags to apply to the item. TagList consists of a list of tag names, which replace any existing tags for the item. TagList may be an empty list. 

5. width=lineLength 
Specifies a maximum line length for the text. If this option is zero (the default) the text is broken into lines only at newline characters. However, if this option is non-zero then any line that would be longer than lineLength is broken just before a space character to make the line shorter than lineLength; the space character is treated as if it were a newline character.