Create ch6_prog_10_strange_fonts_text_1.py
Objective: Display unusual and creative fonts. 

Comments: The value of defining font specification parameters via a function definition is that any of the parameter can become an effective for the duration when the program is run - they whole group of parameters not not need to re-declared every time new text is displayed.